requirement.txt
Django==1.8.2
django-filter==0.10.0
djangorestframework==3.1.3
djangorestframework-bulk==0.2.1
Markdown==2.6.2
